refactor: rename lockfile-converter
