+++ scopes/dependencies/pnpm/lockfile-deps-graph-converter.spec.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 import { DependenciesGraph } from '@teambit/legacy/dist/scope/models/dependencies-graph';
 import { type LockfileFileV9 } from '@pnpm/lockfile.types';
-import { convertLockfileToGraph, convertGraphToLockfile } from './lockfile-converter';
+import { convertLockfileToGraph, convertGraphToLockfile } from './lockfile-deps-graph-converter';
 import { expect } from 'chai';
 
 describe('convertLockfileToGraph simple case', () => {
