+++ scopes/dependencies/pnpm/lockfile-deps-graph-converter.ts
